+++ src/Breadcrumbs.php
@@ -40,7 +40,7 @@ class Breadcrumbs extends Widget
      */
     public $encodeLabels = true;
     /**
-     * @var array the first hyperlink in the breadcrumbs (called home link).
+     * @var array|bool the first hyperlink in the breadcrumbs (called home link).
      * Please refer to [[links]] on the format of the link.
      * If this property is not set, it will default to a link pointing to [[\yii\web\Application::homeUrl]]
      * with the label 'Home'. If this property is false, the home link will not be rendered.
@@ -113,7 +113,7 @@ class Breadcrumbs extends Widget
                 'label' => 'Home',
                 'url' => '/',
             ], $this->itemTemplate);
-        } else {
+        } elseif ($this->homeLink !== false) {
             $links[] = $this->renderItem($this->homeLink, $this->itemTemplate);
         }
 
@@ -165,11 +165,11 @@ class Breadcrumbs extends Widget
      * If this property is not set, it will default to a link pointing with the label 'Home'. If this property is false,
      * the home link will not be rendered.
      *
-     * @param array $value
+     * @param array|bool $value
      *
      * @return $this
      */
-    public function homeLink(array $value): self
+    public function homeLink($value): self
     {
         $this->homeLink = $value;
 
@@ -295,5 +295,4 @@ class Breadcrumbs extends Widget
 
         return strtr($template, ['{link}' => $linkHtml]);
     }
-
 }